# specs/001-fleet-roster-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Pickup MaaS API - 차량 및 승객명단 관리
  description: |
    B2B 송영 서비스를 위한 RESTful API

    **Features**:
    - 001번: 기관 차량 및 승객명단 관리
    - 002번: 주간보호 8시간 케어 스케줄 검증

    **Authentication**: Bearer JWT token required for all endpoints

    **API Conventions**:
    - All timestamps in ISO 8601 format (UTC)
    - All responses include standard error format
    - Pagination uses offset-based strategy
    - Phone numbers stored without hyphens (01012345678)
    - Time values in HH:MM format (24-hour)

  version: 1.0.0
  contact:
    name: Pickup MaaS Team
    email: dev@pickup-maas.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api-staging.pickup-maas.com/v1
    description: Staging environment
  - url: https://api.pickup-maas.com/v1
    description: Production environment
...
security:
  - BearerAuth: []
paths:
...
  /institutions/{institutionId}/passengers:
    get:
      tags: [Passengers]
      summary: 승객 목록 조회
      description: FR-007 - 기관의 모든 승객 조회 (페이지네이션, 필터링)
      operationId: getPassengers
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: shuttleType
          in: query
          description: 셔틀 유형 필터
          schema:
            $ref: '#/components/schemas/ShuttleType'
        - name: groupId
          in: query
          description: 그룹별 필터
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: 이름 또는 연락처 검색
          schema:
            type: string
      responses:
        '200':
          description: 승객 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPassengers'

    post:
      tags: [Passengers]
      summary: 승객 등록
      description: FR-001 - 신규 승객 등록
      operationId: createPassenger
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePassengerDto'
      responses:
        '201':
          description: 승객 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
        '409':
          description: 연락처 중복
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /institutions/{institutionId}/passengers/bulk-upload:
    post:
      tags: [Passengers]
      summary: 승객 일괄 업로드 (엑셀)
      description: FR-015 - 엑셀 파일로 승객 정보 일괄 등록
      operationId: bulkUploadPassengers
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel file (.xlsx)
      responses:
        '201':
          description: 일괄 업로드 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkUploadResult'
        '400':
          description: 엑셀 파일 검증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkUploadError'
  /institutions/{institutionId}/passengers/{passengerId}:
    get:
      tags: [Passengers]
      summary: 승객 상세 조회
      description: FR-008 - 승객 정보 및 스케줄 조회
      operationId: getPassenger
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PassengerIdParam'
      responses:
        '200':
          description: 승객 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'

    patch:
      tags: [Passengers]
      summary: 승객 정보 수정
      description: FR-002 - 승객 정보 수정
      operationId: updatePassenger
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PassengerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePassengerDto'
      responses:
        '200':
          description: 승객 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'

    delete:
      tags: [Passengers]
      summary: 승객 삭제
      description: FR-003 - 승객 삭제 (CASCADE - 스케줄도 삭제됨)
      operationId: deletePassenger
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PassengerIdParam'
      responses:
        '204':
          description: 승객 삭제 성공
  /institutions/{institutionId}/passengers/{passengerId}/schedule:
    put:
      tags: [Passengers]
      summary: 승객 스케줄 설정/수정
      description: 002-FR-001 - 승객의 희망 탑승/하차 시간 입력
      operationId: upsertPassengerSchedule
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PassengerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PassengerScheduleDto'
      responses:
        '200':
          description: 스케줄 저장 성공 (케어 시간 자동 계산됨)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassengerSchedule'

    delete:
      tags: [Passengers]
      summary: 승객 스케줄 삭제
      operationId: deletePassengerSchedule
      parameters:
        - $ref: '#/components/parameters/InstitutionIdParam'
        - $ref: '#/components/parameters/PassengerIdParam'
      responses:
        '204':
          description: 스케줄 삭제 성공

  # ============================================================================
  # Care Time Report Endpoints (002번 기능)
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    InstitutionIdParam:
      name: institutionId
      in: path
      required: true
      description: 기관 ID (UUID)
      schema:
        type: string
        format: uuid
...
    PassengerIdParam:
      name: passengerId
      in: path
      required: true
      description: 승객 ID (UUID)
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  schemas:
...
    PassengerGroup:
      type: object
      required:
        - id
        - institutionId
        - groupCode
        - name
        - totalPassengerCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        institutionId:
          type: string
          format: uuid
        groupCode:
          type: string
          maxLength: 20
          description: 그룹 코드 (기관 내 고유)
          example: 'GRP-001'
        name:
          type: string
          maxLength: 100
          description: 그룹명
          example: '오전 A조'
        totalPassengerCount:
          type: integer
          description: 총 승객 수
          example: 15
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ShuttleType:
      type: string
      enum: [MORNING, EVENING, TEMPORARY]
      description: |
        셔틀 유형:
        - MORNING: 오전 셔틀 (여러 탑승지 → 기관)
        - EVENING: 오후 셔틀 (기관 → 여러 하차지)
        - TEMPORARY: 임시 셔틀 (동적 요청)
    Passenger:
      type: object
      required:
        - id
        - name
        - phoneNumber
        - pickupAddress
        - dropoffAddress
        - shuttleType
        - institutionId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
          example: '김철수'
        phoneNumber:
          type: string
          pattern: '^01[0-9]{9}$'
          description: 연락처 (하이픈 제거)
          example: '01012345678'
        pickupAddress:
          type: string
          maxLength: 200
          example: '서울특별시 강남구 테헤란로 123'
        dropoffAddress:
          type: string
          maxLength: 200
          example: '서울특별시 강남구 역삼동 456'
        shuttleType:
          $ref: '#/components/schemas/ShuttleType'
        institutionId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
          nullable: true
        group:
          $ref: '#/components/schemas/PassengerGroup'
        schedule:
          $ref: '#/components/schemas/PassengerSchedule'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePassengerDto:
      type: object
      required:
        - name
        - phoneNumber
        - pickupAddress
        - dropoffAddress
        - shuttleType
      properties:
        name:
          type: string
          maxLength: 50
          example: '김철수'
        phoneNumber:
          type: string
          pattern: '^01[0-9]-?[0-9]{3,4}-?[0-9]{4}$'
          description: 연락처 (하이픈 포함 가능, 저장 시 제거됨)
          example: '010-1234-5678'
        pickupAddress:
          type: string
          maxLength: 200
        dropoffAddress:
          type: string
          maxLength: 200
        shuttleType:
          $ref: '#/components/schemas/ShuttleType'
        groupId:
          type: string
          format: uuid
          nullable: true
        schedule:
          $ref: '#/components/schemas/PassengerScheduleDto'
    UpdatePassengerDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
        phoneNumber:
          type: string
          pattern: '^01[0-9]-?[0-9]{3,4}-?[0-9]{4}$'
        pickupAddress:
          type: string
          maxLength: 200
        dropoffAddress:
          type: string
          maxLength: 200
        shuttleType:
          $ref: '#/components/schemas/ShuttleType'
        groupId:
          type: string
          format: uuid
          nullable: true
    PaginatedPassengers:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - totalPages
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Passenger'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    BulkUploadResult:
      type: object
      required:
        - successCount
        - failureCount
        - totalRows
        - errors
      properties:
        successCount:
          type: integer
          description: 성공한 행 수
        failureCount:
          type: integer
          description: 실패한 행 수
        totalRows:
          type: integer
          description: 전체 행 수 (헤더 제외)
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BulkUploadRowError'
    BulkUploadRowError:
      type: object
      required:
        - row
        - errors
      properties:
        row:
          type: integer
          description: 에러가 발생한 행 번호
        data:
          type: object
          description: 해당 행의 원본 데이터
        errors:
          type: array
          items:
            type: string
          description: 검증 에러 메시지 목록
    BulkUploadError:
      type: object
      required:
        - statusCode
        - message
        - errors
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: '엑셀 파일 검증에 실패했습니다'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BulkUploadRowError'

    # ========================================================================
    # Passenger Schedule Schemas (002번 기능)
    # ========================================================================
    PassengerSchedule:
      type: object
      required:
        - id
        - passengerId
        - desiredPickupTime
        - desiredDropoffTime
        - careTimeHours
        - isCareTimeInsufficient
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        passengerId:
          type: string
          format: uuid
        desiredPickupTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: 희망 탑승 시간 (HH:MM)
          example: '09:00'
        desiredDropoffTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          description: 희망 하차 시간 (HH:MM)
          example: '17:30'
        careTimeHours:
          type: number
          format: float
          description: 케어 시간 간격 (시간 단위, 자동 계산)
          example: 8.5
        isCareTimeInsufficient:
          type: boolean
          description: 8시간 미만 여부 (자동 계산)
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PassengerScheduleDto:
      type: object
      required:
        - desiredPickupTime
        - desiredDropoffTime
      properties:
        desiredPickupTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          example: '09:00'
        desiredDropoffTime:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
          example: '17:30'
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: '요청 처리 중 오류가 발생했습니다'
        error:
          type: string
          example: 'Bad Request'
        details:
          type: array
          items:
            type: string
          description: 상세 에러 메시지 목록
...